Fixes (dirty) the problem described in #6
I'm not sure how this was supposed to work, but I fixed the link problems (issue #6) so that year and month archives work again. I guess the system already looks for the non existing archive page, not the currently used taxonomy pages, to make the links work. As this page does not exist yet this is a quick and dirty fix for it. 
I would love to make the "show more" button work, but I don't know what the functionality should look like - details welcome. ;)